finish filter data.

+++ src/Controller/MethodsController.php
@@ -14,6 +14,7 @@ class MethodsController extends AppController
 	public function index()
 	{
 		if ($this->request->is('post')) {
+			$auth = $this->Auth->user();
 			$request_body = $this->request->input('json_decode');
 			$data = (array)$request_body;
 			$condition = [];
@@ -26,10 +27,25 @@ class MethodsController extends AppController
 					$condition['Methods.display ILIKE '] = "%$keywords%";
 				}
 			}
+			$super_user = $this->Response->allowOnlySuperUser($auth['group_id']);
 			$query = $this->Methods->find()
 						->contain(['Modules'])
-						->where($condition)
-						->order(['Methods.sort' => 'asc']);
+						->where($condition);
+			if (!$super_user) {
+				$role = $this->Response->getFilterRoleByGroup($auth['group_id']);
+				if ($role) {
+					$query->matching('Permissions', function($permission) use ($role) {
+						return $permission->where([
+							'Permissions.role_id' => $role['role_id'],
+						]);
+					})
+					->order(['Methods.sort' => 'asc']);
+				} else {
+					$query = false;
+				}
+			} else {
+				$query->order(['Methods.sort' => 'asc']);
+			}
 			$data = [];
 			if ($query) {
 				$data = $query->toArray();
@@ -43,19 +59,27 @@ class MethodsController extends AppController
 	public function add()
 	{
 		if ($this->request->is('post')) {
-			$method = $this->Methods->newEntity();
-			$request_body = $this->request->input('json_decode');
-			$data = (array)$request_body;
-			$entity = $this->Methods->patchEntity($method, $data);
-			if ($this->Methods->save($entity)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$method = $this->Methods->newEntity();
+				$request_body = $this->request->input('json_decode');
+				$data = (array)$request_body;
+				$entity = $this->Methods->patchEntity($method, $data);
+				if ($this->Methods->save($entity)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Add not success';
+					return $this->Response->Response($http_code, $message, null, $entity->errors());
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Add not success';
+				$http_code = 403;
+				$message = 'Unauthorized';
 				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
+			
 		}
 	}
 
@@ -81,17 +105,24 @@ class MethodsController extends AppController
 	public function edit()
 	{
 		if ($this->request->is(['patch', 'post', 'put'])) {
-			$request_body = $this->request->input('json_decode');
-			$method = $this->Methods->get($request_body->id);
-			$data = (array)$request_body;
-			$entity = $this->Methods->patchEntity($method, $data);
-			if ($this->Methods->save($entity)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$request_body = $this->request->input('json_decode');
+				$method = $this->Methods->get($request_body->id);
+				$data = (array)$request_body;
+				$entity = $this->Methods->patchEntity($method, $data);
+				if ($this->Methods->save($entity)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Edit not success';
+					return $this->Response->Response($http_code, $message, null, $entity->errors());
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Edit not success';
+				$http_code = 403;
+				$message = 'Unauthorized';
 				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
 		}
@@ -100,16 +131,23 @@ class MethodsController extends AppController
 	public function delete()
 	{
 		if ($this->request->is(['patch', 'post', 'put'])) {
-			$request_body = $this->request->input('json_decode');
-			$method = $this->Methods->get($request_body->id);
-			if ($this->Methods->delete($method)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$request_body = $this->request->input('json_decode');
+				$method = $this->Methods->get($request_body->id);
+				if ($this->Methods->delete($method)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Delete not success';
+					return $this->Response->Response($http_code, $message, null, null);
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Delete not success';
-				return $this->Response->Response($http_code, $message, null, null);
+				$http_code = 403;
+				$message = 'Unauthorized';
+				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
 		}
 	}
