+++ src/Controller/ModulesController.php
@@ -14,9 +14,36 @@ class ModulesController extends AppController
 	public function lists()
 	{
 		if ($this->request->is('post')) {
-			$query = $this->Modules->find()
-			->order(['Modules.sort', 'Modules.name'])
-			->contain(['Methods']);
+			$auth = $this->Auth->user();
+			$query = $this->Modules->find();
+			$super_user = $this->Response->allowOnlySuperUser($auth['group_id']);
+			if (!$super_user) {
+				$role = $this->Response->getFilterRoleByGroup($auth['group_id']);
+				if ($role) {
+					$query->innerJoinWith('Permissions', function($permission) use ($role) {
+						return $permission
+							->where([
+								'Permissions.role_id' => $role['role_id'],
+							]);
+					})
+					->contain('Methods', function ($method) use ($role) {
+						return $method->innerJoinWith('Permissions', function($method_permission) use ($role) {
+							return $method_permission
+								->where([
+									'Permissions.role_id' => $role['role_id'],
+								]);
+						});
+					})
+					->group(['Modules.id'])
+					->order(['Modules.sort' => 'asc']);
+				} else {
+					$query = false;
+				}
+			} else {
+				$query
+					->contain(['Methods'])
+					->order(['Modules.sort' => 'asc']);
+			}
 			$data = [];
 			if ($query) {
 				$data = $query;
@@ -30,6 +57,7 @@ class ModulesController extends AppController
 	public function index()
 	{
 		if ($this->request->is('post')) {
+			$auth = $this->Auth->user();
 			$request_body = $this->request->input('json_decode');
 			$data = (array)$request_body;
 			$condition = [];
@@ -39,10 +67,37 @@ class ModulesController extends AppController
 					$condition['Modules.name ILIKE '] = "%$keywords%";
 				}
 			}
-			$query = $this->Modules->find()
-						->contain(['Methods'])
-						->where($condition)
-						->order(['Modules.sort' => 'asc']);
+			$query = $this->Modules->find();
+			$super_user = $this->Response->allowOnlySuperUser($auth['group_id']);
+			if (!$super_user) {
+				$role = $this->Response->getFilterRoleByGroup($auth['group_id']);
+				if ($role) {
+					$query->innerJoinWith('Permissions', function($permission) use ($role) {
+						return $permission
+							->where([
+								'Permissions.role_id' => $role['role_id'],
+							]);
+					})
+					->contain('Methods', function ($method) use ($role) {
+						return $method->innerJoinWith('Permissions', function($method_permission) use ($role) {
+							return $method_permission
+								->where([
+									'Permissions.role_id' => $role['role_id'],
+								]);
+						});
+					})
+					->where($condition)
+					->group(['Modules.id'])
+					->order(['Modules.sort' => 'asc']);
+				} else {
+					$query = false;
+				}
+			} else {
+				$query
+					->contain(['Methods'])
+					->where($condition)
+					->order(['Modules.sort' => 'asc']);
+			}
 			$data = [];
 			if ($query) {
 				$data = $query;
@@ -56,19 +111,27 @@ class ModulesController extends AppController
 	public function add()
 	{
 		if ($this->request->is('post')) {
-			$module = $this->Modules->newEntity();
-			$request_body = $this->request->input('json_decode');
-			$data = (array)$request_body;
-			$entity = $this->Modules->patchEntity($module, $data);
-			if ($this->Modules->save($entity)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$module = $this->Modules->newEntity();
+				$request_body = $this->request->input('json_decode');
+				$data = (array)$request_body;
+				$entity = $this->Modules->patchEntity($module, $data);
+				if ($this->Modules->save($entity)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Add not success';
+					return $this->Response->Response($http_code, $message, null, $entity->errors());
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Add not success';
+				$http_code = 403;
+				$message = 'Unauthorized';
 				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
+			
 		}
 	}
 
@@ -86,7 +149,6 @@ class ModulesController extends AppController
 				$message = 'User not found.';
 				return $this->Response->Response($http_code, $message, null, null);
 			}
-			
 		}
 		
 	}
@@ -94,17 +156,24 @@ class ModulesController extends AppController
 	public function edit()
 	{
 		if ($this->request->is(['patch', 'post', 'put'])) {
-			$request_body = $this->request->input('json_decode');
-			$module = $this->Modules->get($request_body->id);
-			$data = (array)$request_body;
-			$entity = $this->Modules->patchEntity($module, $data);
-			if ($this->Modules->save($entity)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$request_body = $this->request->input('json_decode');
+				$module = $this->Modules->get($request_body->id);
+				$data = (array)$request_body;
+				$entity = $this->Modules->patchEntity($module, $data);
+				if ($this->Modules->save($entity)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Edit not success';
+					return $this->Response->Response($http_code, $message, null, $entity->errors());
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Edit not success';
+				$http_code = 403;
+				$message = 'Unauthorized';
 				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
 		}
@@ -113,16 +182,23 @@ class ModulesController extends AppController
 	public function delete()
 	{
 		if ($this->request->is(['patch', 'post', 'put'])) {
-			$request_body = $this->request->input('json_decode');
-			$module = $this->Modules->get($request_body->id);
-			if ($this->Modules->delete($module)) {
-				$http_code = 200;
-				$message = 'Success';
-				return $this->Response->Response($http_code, $message);
+			$auth = $this->Auth->user();
+			if ($this->Response->allowOnlySuperUser($auth['group_id'])) {
+				$request_body = $this->request->input('json_decode');
+				$module = $this->Modules->get($request_body->id);
+				if ($this->Modules->delete($module)) {
+					$http_code = 200;
+					$message = 'Success';
+					return $this->Response->Response($http_code, $message);
+				} else {
+					$http_code = 400;
+					$message = 'Delete not success';
+					return $this->Response->Response($http_code, $message, null, null);
+				}
 			} else {
-				$http_code = 400;
-				$message = 'Delete not success';
-				return $this->Response->Response($http_code, $message, null, null);
+				$http_code = 403;
+				$message = 'Unauthorized';
+				return $this->Response->Response($http_code, $message, null, $entity->errors());
 			}
 		}
 	}
